+++ .dotnet/src/Custom/Chat/ChatResponseFormat.cs
@@ -1,68 +1,68 @@
-using System;
-using System.ComponentModel;
 using OpenAI.Internal;
+using System;
+using System.ClientModel.Primitives;
+using System.ComponentModel;
 
 namespace OpenAI.Chat;
 
-/// <summary>
-/// Represents a requested <c>response_format</c> for the model to use, enabling "JSON mode" for guaranteed valid output.
-/// </summary>
-/// <remarks>
-/// <b>Important</b>: when using JSON mode, the model <b><u>must</u></b> also be instructed to produce JSON via a
-/// <c>system</c> or <c>user</c> message.
-/// <para>
-/// Without this paired, message-based accompaniment, the model may generate an unending stream of whitespace until the
-/// generation reaches the token limit, resulting in a long-running and seemingly "stuck" request.
-/// </para>
-/// <para>
-/// Also note that the message content may be partially cut off if <c>finish_reason</c> is <c>length</c>, which
-/// indicates that the generation exceeded <c>max_tokens</c> or the conversation exceeded the max context length for
-/// the model.
-/// </para>
-/// </remarks>
 [CodeGenModel("ChatResponseFormat")]
 public abstract partial class ChatResponseFormat
 {
-    /// <summary> text. </summary>
     public static ChatResponseFormat Text { get; } = new InternalChatResponseFormatText();
-    /// <summary> json_object. </summary>
     public static ChatResponseFormat JsonObject { get; } = new InternalChatResponseFormatJsonObject();
 
-    public static ChatResponseFormat FromJsonSchema(
+    public static ChatResponseFormat CreateTextFormat() => new InternalChatResponseFormatText();
+    public static ChatResponseFormat CreateJsonObjectFormat() => new InternalChatResponseFormatJsonObject();
+    public static ChatResponseFormat CreateJsonSchemaFormat(
         string name,
         string description = null,
         BinaryData jsonSchema = null,
-        bool? useStrictSchema = null)
+        bool? requireStrictJsonSchemaMatch = null)
     {
         InternalResponseFormatJsonSchemaJsonSchema internalSchema = new(
             description,
             name,
             jsonSchema,
-            useStrictSchema,
+            requireStrictJsonSchemaMatch,
             null);
         return new InternalChatResponseFormatJsonSchema(internalSchema);
     }
 
     [EditorBrowsable(EditorBrowsableState.Never)]
+    public static bool operator ==(ChatResponseFormat first, ChatResponseFormat second)
+        => first?.Equals(second) ?? false;
+
+    [EditorBrowsable(EditorBrowsableState.Never)]
+    public static bool operator !=(ChatResponseFormat first, ChatResponseFormat second)
+        => !(first == second);
+
+    [EditorBrowsable(EditorBrowsableState.Never)]
     public override bool Equals(object obj)
     {
-        return
-            (this is InternalChatResponseFormatText && obj is InternalChatResponseFormatText)
-            || (this is InternalChatResponseFormatJsonObject && obj is InternalChatResponseFormatJsonObject)
-            || (this is InternalChatResponseFormatJsonSchema thisSchema && obj is InternalChatResponseFormatJsonSchema otherSchema && thisSchema.JsonSchema.Name == otherSchema.JsonSchema.Name);
+        if (this is InternalChatResponseFormatText)
+        {
+            return obj is InternalChatResponseFormatText;
+        }
+        else if (this is InternalChatResponseFormatJsonObject)
+        {
+            return obj is InternalChatResponseFormatJsonObject;
+        }
+        else if (this is InternalChatResponseFormatJsonSchema thisJsonSchema
+            && obj is InternalChatResponseFormatJsonSchema otherJsonSchema)
+        {
+            return thisJsonSchema.JsonSchema.Name.Equals(otherJsonSchema.JsonSchema.Name);
+        }
+        else
+        {
+            return ToString().Equals(obj?.ToString());
+        }
     }
 
     [EditorBrowsable(EditorBrowsableState.Never)]
-    public static bool operator==(ChatResponseFormat first, ChatResponseFormat second)
-        => first.Equals(second);
+    public override int GetHashCode() => ToString().GetHashCode();
 
-    [EditorBrowsable(EditorBrowsableState.Never)]
-    public static bool operator!=(ChatResponseFormat first, ChatResponseFormat second)
-        => !first.Equals(second);
-
-    [EditorBrowsable(EditorBrowsableState.Never)]
-    public override int GetHashCode()
+    public override string ToString()
     {
-        return base.GetHashCode();
+        return ModelReaderWriter.Write(this).ToString();
     }
 }